+++ src/modules/icmaa-cms/pages/LandingPage.vue
@@ -1,0 +1,20 @@
+<template>
+  <div id="cms-page" class="t-container t-p-4">
+    Hallo welt
+  </div>
+</template>
+
+<script>
+import CmsMetaMixin from 'icmaa-meta/mixins/cmsMeta'
+
+export default {
+  name: 'LandingPage',
+  mixins: [ CmsMetaMixin ],
+  components: {
+
+  },
+  props: {
+
+  }
+}
+</script>
