+++ src/themes/icmaa-imp/components/core/blocks/ICMAA/Cms/Pages/Affiliate.vue
@@ -68,7 +68,7 @@
 </template>
 
 <script>
-import Page from 'icmaa-cms/components/Page'
+import Page from 'icmaa-cms/mixins/Page'
 import MaterialIcon from 'theme/components/core/blocks/MaterialIcon'
 
 export default {
