+++ src/themes/icmaa-imp/components/core/blocks/ICMAA/Cms/Pages/ImpericonNeverSayDie.vue
@@ -140,7 +140,7 @@
 </template>
 
 <script>
-import Page from 'icmaa-cms/components/Page'
+import Page from 'icmaa-cms/mixins/Page'
 
 export default {
   name: 'ImpericonNeverSayDie',
